remove 100AU disc restriction on stars
I don't know if this was an arbitrary choice or if there's a theoretical upper
limit on the size that we should be enforcing, and I don't have the internet
available right now to check. What I do know is that touring the galaxy with
LSHIFT down for a couple of minutes only triggered this condition half a dozen
times, and the largest disc I saw was ~115AU. The difference seems small
enough to not be worth the additional code complexity.