remove requirement that secondary stars in a multiple system be less massive than their partners/primary
Another internet-free executive decision. Intuitively I don't see why the
secondary star should be smaller than primary because the concepts of
"primary" and "secondary" are entirely arbitrary - it depends on your point of
reference.

Long ago (added in 5ab5761d, removed in 4f3298b7), the sysgen had this comment
as a kind of justification:

    // normally star types are picked by spectral class distribution in
    // our galactic neighbourhood. in binary systems instead just pick
    // random companion types up to spectral class of primary.

So fair enough - the "primary" should be visually dominant, because otherwise
we'd see it as a different colour from afar. However, this is implemented (by
type, not mass) in Sector.cpp. I will make the argument that the sysgen should
generate whateever it is asked to by the SystemDescriptor. If it gets
something that makes no sense then that's the fault of whatever created the
descriptor (probably some custom system) and its not for us to blithely
override it.